v0.03 - Creation of translated files
This time we started to create the file structures to translate the program into some languages, they are (Portuguese, English, Spanish, French, Italian, German, Chinese, Japanese, Russian). We also created some classes that will be used in our project, classes that we will use to manipulate the program and the user's operating system.